add a "normalized" file for basic_qualifications
Many job ads have minor differences in some of their attributes. One of
the easiest, useful attributes for me to normalize* on is the
basic_qualifications attribute, and so I am focusing on that. To that
end, I created a function to process a file containing the
basic_qualifications attribute of a job ad. It's not perfect, and I
haven't verified thoroughly yet. I originally started writing the
function in vim, and there were many things to change in the
translation, and one thing that I haven't been able to translate yet
(involves lazy/reluctant/non-greedy searches), but it seemed to be
working on a couple of job ads I inspected.

I hoping to proceed with composing some measure for similarity of these
normalized files, and maybe group them together. Exact md5sum hashes
would be the simplest, for exact matches. Next order up would be lengths
of the diffs between sorted versions of the files (or maybe I should add
sorting to the normalized versions of the files...).

* Replace minor differences with a common version. This invovles a lot
of substituations, with order mattering for some of them.